Exclude doc target from all
In order to be able to better separate the generation of the
documentation from the build process itself, the doc target is excluded
from the all (or default) target. This implies that for a successful
`package` target run, in case of a configured build to generate the
documentation (`-DCQLITE_BUILD_DOCUMENTATION=ON`), the `doc` target has
to be run prior to calling the `package` or `install` target.