Fixed build system pollution
In order to play better as a subproject, all "PACKAGE_" names have been
renamed to "CQLITE_" in order to not disturb the packaging of the
including package.

There was a bug in using the git command when a squash merge did not
contain a version for git describe to list. The git version is only
taken into account when this is not the case, otherwise the
corresponding symbols (CQLITE_GIT_PROJECT_VERSION and
CQLITE_GIT_COMMID_ID) are undefined.