Use find_package for SQLite, do cleanups
Instead of searching for the `sqlite` library use the existing `find_package`.
Cleanup the `cmake` files and remove the leftover global include directory. Do
not attempt to install any `.dll` files for now. The current approach would not
work, since there is a `.lib` file for every `.dll` file that would as well
need to be included.